requirements.txt
Flask
python-dotenv
Pillow
pytesseract
google-generativeai
mistralai
huggingface_hub
openai